requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiobotocore==2.21.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.16
aioitertools==0.12.0
aiosignal==1.2.0
alabaster==1.0.0
annotated-types==0.7.0
antlr4-python3-runtime==4.13.2
astunparse==1.6.3
attrs==24.3.0
aws-sam-translator==1.97.0
aws-xray-sdk==2.14.0
babel==2.17.0
beautifulsoup4==4.13.4
blinker==1.9.0
blis==1.3.0
boto3==1.37.1
botocore==1.37.1
botok==0.9.0
Brotli==1.0.9
catalogue==2.0.10
certifi==2025.4.26
cffi==1.17.1
cfn-lint==1.34.2
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.21.0
confection==0.1.5
courlan==1.3.2
cryptography==44.0.2
cymem==2.0.11
datasets==3.5.1
dateparser==1.2.1
dill==0.3.8
docker==7.1.0
docutils==0.21.2
emoji==2.14.1
execnet==2.1.1
fasteners==0.19
fasttext-numpy2-wheel==0.9.2
fasttext-wheel==0.9.2
faust-cchardet==2.1.19
filelock==3.13.1
Flask==3.1.0
flask-cors==5.0.1
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
gast==0.6.0
google-pasta==0.2.0
graphql-core==3.2.6
grpcio==1.71.0
h5py==3.13.0
hf-xet==1.1.0
htmldate==1.9.3
huggingface-hub==0.30.2
humanize==3.10.0
idna==3.7
imagesize==1.4.1
indic-nlp-library==0.92
iniconfig==2.1.0
inscriptis==2.6.0
itsdangerous==2.2.0
jieba==0.42.1
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
joserfc==1.0.4
jsonpatch==1.33
jsonpath-ng==1.7.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-path==0.3.4
jsonschema-specifications==2025.4.1
jusText==3.0.2
kenlm==0.3.0
keras==3.9.2
khmer-nltk==1.6
kiwipiepy==0.20.4
kiwipiepy_model==0.20.0
langcodes==3.5.0
language_data==1.3.0
LaoNLP==1.2.0
lazy-object-proxy==1.11.0
libclang==18.1.1
loguru==0.7.2
lxml==5.4.0
lxml_html_clean==0.4.2
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==2.2.0
MarkupSafe==3.0.2
mdurl==0.1.0
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
ml_dtypes==0.5.1
Morfessor==2.0.6
moto==5.1.4
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
murmurhash==1.0.12
namex==0.0.9
networkx==3.4.2
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.1
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pathable==0.4.4
pillow==11.3.0
pip==25.0
pluggy==1.5.0
ply==3.11
preshed==3.0.9
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
py-partiql-parser==0.6.1
pyahocorasick==2.1.0
pyarrow==20.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pyidaungsu-numpy2==0.1.4
pyparsing==3.2.3
PySocks==1.7.1
pytest==8.3.5
pytest-timeout==2.3.1
pytest-xdist==3.6.1
pythainlp==5.1.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-magic==0.4.27
pytz==2025.2
pyvi==0.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
responses==0.25.7
rfc3339-validator==0.1.4
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.24.0
ruff==0.11.7
s3fs==2025.3.0
s3transfer==0.11.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
setuptools==72.1.0
shellingham==1.5.4
six==1.17.0
sklearn-crfsuite==0.5.0
smart-open==7.1.0
snowballstemmer==2.2.0
soupsieve==2.7
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
srsly==2.5.1
stanza==1.10.1
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
thinc==8.3.6
threadpoolctl==3.6.0
tld==0.13
tldextract==5.3.0
tokenizers==0.20.3
torch==2.7.0
tqdm==4.67.1
trafilatura==1.11.0
transformers==4.45.1
triton==3.3.0
typer==0.15.2
typing_extensions==4.12.2
typing-inspection==0.4.0
tzdata==2025.2
tzlocal==5.3.1
urduhack==0.1.4
uritools==4.0.3
urllib3==2.3.0
warcio==1.7.5
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
xmltodict==0.14.2
xxhash==3.5.0
yarl==1.18.0
zstandard==0.23.0
